pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gh-release-devpi"
version = "0.1.0"
description = "GitHub Release Artifact Downloader and DevPI Uploader"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygithub>=2.1.1",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.scripts]
gh-release-devpi = "gh_release_devpi.main:app"
